Set ERC20 approval on load
Includes a small refactor to useERC20Events handler
